mcp-server/pyproject.toml
[project]
name = "hellio-hr-templates-mcp"
version = "1.0.0"
description = "MCP server for HR document template management"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.5.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = []
